requirements-dev.txt
-r requirements.txt
pandas-stubs>=2.2.0.240218
flake8>=7.1.0
